Make all private methods and properties protected, change internal naming scheme.
For better extensibility, all "private" members are now protected,
enabling them to be called from subclasses. We are all
consenting adults - if you want to poke around in Idiorm's internals, be
my guest. Additionally, all protected members and methods now start with
an underscore to make them more visible in the code. This approach is
based on Kohana's public/private and naming policies. Closes #4.