querying instances of extended classes
if idiorm is used with paris (or other classes that extend ORM) querying should return instances of the class that extends ORM instead of ORM.
$user = Model::factory('User')->find_one($id);
and
$user = Model::factory('User')->create();
should both have ORMWrapper as $orm property. if self:: instead of static is used find_one will set instance of ORM to the §orm property of the model instead of ORMWrapper.